% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_node_type}
\alias{add_node_type}
\title{Adding genetype to the gene names to distinguish biological function}
\usage{
add_node_type(df)
}
\arguments{
\item{df}{dataframe with interaction data}
}
\value{
df
}
\description{
Adding genetype to the gene names to distinguish biological function
}
