% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enrich}
\alias{enrich}
\title{Ranking the most interactive gene (ligand or receptor)}
\usage{
enrich(list, name, org = org.Hs.eg.db, univ = NULL)
}
\arguments{
\item{list}{of genes}

\item{rankings's}{table column name}
}
\value{
list
}
\description{
Ranking the most interactive gene (ligand or receptor)
}
