% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filtered_graphs}
\alias{filtered_graphs}
\title{Filter comparison graphs by statistics}
\usage{
filtered_graphs(data, out_path)
}
\arguments{
\item{data}{datafromlian}

\item{out_path}{save path}
}
\value{
tibble
}
\description{
Filter comparison graphs by statistics
}
