% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fisher_test_cci}
\alias{fisher_test_cci}
\title{Evaluate Differences in the edge proportion}
\usage{
fisher_test_cci(data, measure, out_path)
}
\arguments{
\item{data}{datafromlian}

\item{measure}{intensity}

\item{out_path}{save path}
}
\value{
tibble
}
\description{
Evaluate Differences in the edge proportion
}
