% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report_liana.R
\name{generate_report_liana}
\alias{generate_report_liana}
\title{Run all LR Downstream analysis}
\usage{
generate_report_liana(
  lianalr,
  genes = NULL,
  out_path,
  threshold = 0,
  colors = NULL,
  out_file = NULL,
  report = TRUE,
  output_fmt = "html_document",
  sel_columns = c("source", "target", "Ligand", "Receptor", "meanlr")
)
}
\arguments{
\item{lianalr}{liana object}

\item{genes}{list of genes to be considered in the sankey plots}

\item{out_path}{output directory path}

\item{threshold}{percentage of edges to be pruned}

\item{colors}{celltypes colorscheme}

\item{out_file}{output file names}

\item{report}{decide if a report is generated or not}

\item{output_fmt}{rmarkdown render output format parameter}

\item{sep}{character used on csv}
}
\value{
Rmarkdown report all objects from each step
}
\description{
This function loads the single conditions LR outputs and return the LR network based analysis.
It assumes that the table present the following columns Ligand, Ligand.Cluster, Receptor,Receptor.Cluster and MeanLR/another
measure
}
\examples{
paths <- c('CTR' = system.file("extdata",
                              "ctr_nils_bm_human.csv",
                              package = "CrossTalkeR"),
          'EXP' = system.file("extdata",
                              "exp_nils_bm_human.csv",
                              package = "CrossTalkeR"))
output =  system.file("extdata", package = "CrossTalkeR")
genes <- c('TGFB1')
data <- generate_report(lrpaths = paths,
                       genes = genes,
                       out_path = paste0(output,'/'),
                       threshold = 0,
                         out_file = "report.html")

}
