% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kegg_annotation}
\alias{kegg_annotation}
\title{Annotate Exclusive LR pairs (ligand or receptor)}
\usage{
kegg_annotation(
  data,
  slot,
  out_path,
  database = org.Hs.eg.db::org.Hs.eg.db,
  org = "hsa"
)
}
\arguments{
\item{data}{lrobject}

\item{slot}{table fields}

\item{database}{complex heatmap database}

\item{org}{organism to be considered}

\item{subslot}{table}
}
\description{
Annotate Exclusive LR pairs (ligand or receptor)
}
