% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{liana2CT}
\alias{liana2CT}
\title{Format liana2CT}
\usage{
liana2CT(data, gene_Ai, gene_Bi, measure)
}
\arguments{
\item{data}{datafromliana}

\item{measure}{measure to be considered}

\item{source}{source cell population}

\item{target}{target cell population}

\item{gene_A}{source gene population}

\item{gene_B}{target gene population}

\item{type_gene_A}{type source gene population}

\item{type_gene_B}{type target gene population}
}
\value{
tibble
}
\description{
Format liana2CT
}
