% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PCbiplot}
\alias{PCbiplot}
\title{This function signed sending and receiving barplot (GGI)}
\usage{
PCbiplot(PC, x = "PC1", y = "PC2")
}
\arguments{
\item{PCA}{LRobject}
}
\value{
R default plot
}
\description{
This function signed sending and receiving barplot (GGI)
}
\examples{
PCbiplot(data@pca[['CTR_ggi']])
}
