% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_bar_rankings}
\alias{plot_bar_rankings}
\title{This function generates the barplot for a given network ranking on the CGI level.
Further, the genes can be filtered by selected gene types to filter the plot.}
\usage{
plot_bar_rankings(
  data_object,
  table_name,
  ranking,
  type = NULL,
  filter_sign = NULL,
  mode = "cci",
  top_num = 10
)
}
\arguments{
\item{data_object}{LRobject with all data}

\item{table_name}{name of the ranking table}

\item{ranking}{name of the network ranking to use}

\item{type}{gene type (L,R,TF, LR/RL, RTF/TFR, LTF/TFL)}

\item{filter_sign}{show all (NULL), only positive (pos), or only negativ (neg) results}

\item{mode}{mode of the ranking (cci, ggi)}

\item{top_num}{number of top genes to show in positive or negative direction}
}
\value{
R default plot
}
\description{
This function generates the barplot for a given network ranking on the CGI level.
Further, the genes can be filtered by selected gene types to filter the plot.
}
