% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_bar_rankings_cci}
\alias{plot_bar_rankings_cci}
\title{This function generates the barplot for a given network ranking on the CCI level}
\usage{
plot_bar_rankings_cci(
  data_object,
  table_name,
  ranking = "pagerank",
  filter_sign = NULL
)
}
\arguments{
\item{data_object}{LRobject with all data}

\item{table_name}{name of the ranking table}

\item{ranking}{name of the network ranking to use}

\item{filter_sign}{show all (NULL), only positive (pos), or only negativ (neg) results}
}
\value{
R default plot
}
\description{
This function generates the barplot for a given network ranking on the CCI level
}
