% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cci}
\alias{plot_cci}
\title{Plot Cell Cell Interaction}
\usage{
plot_cci(
  graph,
  col,
  plt_name,
  coords,
  emax = NULL,
  leg = FALSE,
  low = 25,
  high = 75,
  ignore_alpha = F,
  log = F,
  efactor = 8,
  vfactor = 12
)
}
\arguments{
\item{graph}{Paths of single condition LR data}

\item{col}{Cluster coords according do desired layout default is circle layout}

\item{plt_name}{Plot Name (Title)}

\item{emax}{Max MeanLR across the all inputs, if its not defined, the method going to consider the max find within a sample}

\item{leg}{Set color legend}

\item{low}{Lower threshold: This parameter low and high defines the edges which will be filtered. Edges within the interval \link{low,high} are filtered.}

\item{high}{Higher threshould}

\item{efactor}{edge scale factor}

\item{vfactor}{edge scale factor}
}
\value{
R default plot
}
\description{
This function do a CCI plot
}
