% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_deregulated_pathways}
\alias{plot_deregulated_pathways}
\title{This function is a proxy to the PCA plot in comparative conditions}
\usage{
plot_deregulated_pathways(data_object, name, title = NULL)
}
\arguments{
\item{data_object}{LRobject with all data}

\item{name}{name of the table}
}
\value{
R default plot
}
\description{
This function is a proxy to the PCA plot in comparative conditions
}
