% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ggi}
\alias{plot_ggi}
\title{This function do a ggi plot and higest degree nodes}
\usage{
plot_ggi(graph, color, name)
}
\arguments{
\item{graph}{graph}

\item{color}{cluster color}
}
\value{
R default plot
}
\description{
This function do a ggi plot and higest degree nodes
}
