% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_Heatmap}
\alias{plot_Heatmap}
\title{This function generates the graph adjacency matrix Heatmap}
\usage{
plot_Heatmap(graph, weight)
}
\arguments{
\item{graph}{cell-cell communication graph}

\item{weight}{cell-cell communication graph's slot}
}
\value{
R default plot
}
\description{
This function generates the graph adjacency matrix Heatmap
}
