% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pca}
\alias{plot_pca}
\title{This function is a proxy to the PCA plot}
\usage{
plot_pca(lrobj_tblPCA, curr, dims = c(1, 2), ret = F, ggi = TRUE)
}
\arguments{
\item{curr}{table entry}

\item{dims}{PCA dims}

\item{ret}{return plot}

\item{ggi}{GGI mode}

\item{lrobj_tbl}{LRobject table with all data}
}
\value{
R default plot
}
\description{
This function is a proxy to the PCA plot
}
\examples{
paths <- c(
  "CTR" = system.file("extdata",
    "CTR_LR.csv",
    package = "CrossTalkeR"
  ),
  "EXP" = system.file("extdata",
    "EXP_LR.csv",
    package = "CrossTalkeR"
  )
)
output <- system.file("extdata", package = "CrossTalkeR")
genes <- c("TGFB1")

data <- generate_report(paths,
  genes,
  out_path = paste0(output, "/"),
  threshold = 0,
  out_file = "vignettes_example.html",
  output_fmt = "html_document",
  report = FALSE
)
}
