% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pca_barplor}
\alias{plot_pca_barplor}
\title{This function pca PC1 and PC2 ranking barplot}
\usage{
plot_pca_barplor(pca_l, lrobj_table, pc = 1)
}
\arguments{
\item{pca_l}{LRobject pca_slot}

\item{lrobj_table}{LRobject table slcot}
}
\value{
R default plot
}
\description{
This function pca PC1 and PC2 ranking barplot
}
\examples{
data <- generate_report(paths,genes,'~/Documents/',threshold=0,out_file = 'report.html')
plot_pca(lrobj_tbl = data@tables$EXP_x_CTR,
           target = c("TGFB1"),
           ligand_cluster = NULL,
           receptor_cluster = NULL,
           plt_name = "TGFB1")

}
