% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sankey_tf}
\alias{plot_sankey_tf}
\title{This function selected gene transcription factor interaction related sankey plot}
\usage{
plot_sankey_tf(
  lrobj_tbl,
  target = NULL,
  cluster = NULL,
  target_type = NULL,
  plt_name = NULL,
  threshold = 50
)
}
\arguments{
\item{lrobj_tbl}{LRobject table with all data}

\item{target}{gene}

\item{cluster}{cluster}

\item{target_type}{type of target}

\item{plt_name}{plot title}

\item{threshold}{top_n n value}
}
\value{
R default plot
}
\description{
This function selected gene transcription factor interaction related sankey plot
}
\examples{
paths <- c('CTR' = system.file("extdata",
                              "ctr_nils_bm_human.csv",
                              package = "CrossTalkeR"),
          'EXP' = system.file("extdata",
                              "exp_nils_bm_human.csv",
                              package = "CrossTalkeR"))
output =  system.file("extdata", package = "CrossTalkeR")
genes <- c('TGFB1')

data <- generate_report(paths,
                       genes,
                       out_path=paste0(output,'/'),
                       threshold=0,
                       out_file = 'vignettes_example.html',
                       output_fmt = "html_document",
                       report = FALSE)
}
