% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_signedbar}
\alias{plot_signedbar}
\title{This function signed sending and receiving barplot (CCI)}
\usage{
plot_signedbar(all_data, curr)
}
\arguments{
\item{all_data}{LRobject}

\item{curr}{condition}
}
\value{
R default plot
}
\description{
This function signed sending and receiving barplot (CCI)
}
