% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_Volcano}
\alias{plot_Volcano}
\title{This function generates the enhancedVolcanoplot}
\usage{
plot_Volcano(table, feature_name, pvalcutoff = 0.05, x = "lodds", y = "p")
}
\arguments{
\item{table}{Table with the test Results}

\item{feature_name}{column with names}
}
\value{
R default plot
}
\description{
This function generates the enhancedVolcanoplot
}
