% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotInterInfo}
\alias{plotInterInfo}
\title{This function selected genes correlation}
\usage{
plotInterInfo(data, lrobject, pair, lrslot, assay = "RNA", comp = "EXP_x_CTR")
}
\arguments{
\item{data}{Seurat object}

\item{lrobject}{LRObj object}

\item{pair}{pair of interest}

\item{lrslot}{table from lrobject}

\item{assay}{Seurat assay}
}
\value{
R default plot
}
\description{
This function selected genes correlation
}
