% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ranking_cci}
\alias{ranking_cci}
\title{Ranking the most interactive cell type}
\usage{
ranking_cci(data, out)
}
\arguments{
\item{data}{lrobject}

\item{out}{path to save the lrobject with ranking}
}
\value{
list
}
\description{
Ranking the most interactive cell type
}
