% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ranking_ggi}
\alias{ranking_ggi}
\title{Ranking the most interactive gene (ligand or receptor)}
\usage{
ranking_ggi(data, out_path)
}
\arguments{
\item{data}{lrobject}

\item{out}{path to save the lrobject with ranking}
}
\value{
list
}
\description{
Ranking the most interactive gene (ligand or receptor)
}
