% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ranking_net}
\alias{ranking_net}
\title{Network Ranking method}
\usage{
ranking_net(graph, mode = TRUE)
}
\arguments{
\item{graph}{lrobject}

\item{mode}{is TRUE if is comparive mode}
}
\value{
list
}
\description{
Network Ranking method
}
