% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{read_lr_single_condiction}
\alias{read_lr_single_condiction}
\title{Read single condition tables}
\usage{
read_lr_single_condiction(LRpaths, out_path, sep = ",")
}
\arguments{
\item{LRpaths}{Paths of single condition LR data}

\item{sep}{character used on csv}
}
\value{
LRObject
}
\description{
This function loads the single conditions LR outputs
}
\details{
It assumes that the table present the following columns Ligand, Ligand.Cluster, Receptor,Receptor.Cluster and MeanLR/another
measure
}
