% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_condition.R
\name{read_lr_single_condiction_liana}
\alias{read_lr_single_condiction_liana}
\title{Read single condition tables}
\usage{
read_lr_single_condiction_liana(
  lianalr,
  out_path = "/tmp/",
  colors = NULL,
  sel_columns
)
}
\arguments{
\item{lianalr}{liana object}

\item{out_path}{Path to deposit the results}

\item{colors}{colorlist}

\item{sel_columns}{colorlist}
}
\value{
LRObject
}
\description{
This function loads the single conditions LR outputs and use it to generate
the report data and it`s object
It assumes that the table presents the following columns Ligand, Ligand.Cluster, Receptor,Receptor.Cluster and MeanLR/another
measure
}
