% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_condition.R
\name{read_lr_single_condition}
\alias{read_lr_single_condition}
\title{Read single condition tables}
\usage{
read_lr_single_condition(
  lrpaths,
  sel_columns,
  out_path = "/tmp/",
  sep = ",",
  colors = NULL
)
}
\arguments{
\item{lrpaths}{Named vector with the lrpaths of each output}

\item{sel_columns}{selected columns}

\item{out_path}{Path to deposit the results}

\item{sep}{character used to divide the columns on input file}

\item{colors}{colorlist}
}
\value{
LRObject
}
\description{
This function loads the single conditions LR outputs and use it to generate
the report data and it`s object
It assumes that the table presents the following columns Ligand, Ligand.Cluster, Receptor,Receptor.Cluster and MeanLR/another
measure
}
