% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{remove_dosage}
\alias{remove_dosage}
\title{Count Gene Dosage}
\usage{
remove_dosage(final)
}
\arguments{
\item{list}{Paths of single condition LR data}
}
\value{
list
}
\description{
Count Gene Dosage
}
