% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_LR_TF_analysis.R
\name{run_LR_analysis}
\alias{run_LR_analysis}
\title{Running the liana tool to retrieve LR interactions}
\usage{
run_LR_analysis(
  seurat_object,
  condition_ident,
  annotation_ident,
  organism,
  out_path
)
}
\arguments{
\item{seurat_object}{seurat object with expression data}

\item{condition_ident}{name of the meta data field containing condition information to split the object}

\item{annotation_ident}{name of the meta data field containing the cluster annotation}

\item{organism}{either human or mouse}

\item{out_path}{output path to save the results to}
}
\value{
list of LR interaction tables by condition
}
\description{
This function runs the liana_wrap function to retrieve LR interactions for the given seurat object using the
method from CellPhoneDB and the liana Consensus database.
}
