% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_LR_TF_analysis.R
\name{run_LRTF_analysis}
\alias{run_LRTF_analysis}
\title{Running the LR and LR to TF analysis}
\usage{
run_LRTF_analysis(seurat_object, parameters)
}
\arguments{
\item{seurat_object}{seurat object with expression data}

\item{parameters}{list of arguments that are needed to run the LR and TF analyses}
}
\value{
list of LRTF interaction tables by condition that can be used to run CrossTalkeR
}
\description{
Wrapper to run the LR analysis with liana and the TF analysis with the LR2TF package
}
