% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_LR_TF_analysis.R
\name{run_TF_analysis}
\alias{run_TF_analysis}
\title{Running the LR to TF analysis from the LR2TF package}
\usage{
run_TF_analysis(seurat_object, parameters, LR_results = NA)
}
\arguments{
\item{seurat_object}{seurat object with expression data}

\item{parameters}{list of arguments that are needed to run the TF analysis}

\item{LR_results}{tables with LR interactions by consitions}
}
\value{
list of LRTF interaction tables by condition
}
\description{
This function runs a TF activity analysis with the dorothea tool and combines active TFs with Ligand and Receptors
based on expression and a-priori knowledge from the Omnipath database
}
