% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{set_coords}
\alias{set_coords}
\title{This function assigns coordinates in the sankey plots for the genes on the dataframe}
\usage{
set_coords(df, type)
}
\arguments{
\item{df}{dataframe with genes contained in the sankey plot}

\item{type}{string defining the type of the genes inside the dataframe (L, R or TF)}
}
\description{
This function assigns coordinates in the sankey plots for the genes on the dataframe
}
