% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{comparative_med}
\alias{comparative_med}
\title{Delta betweenness the most interactive gene (ligand or receptor)}
\usage{
comparative_med(rankings, slotname, graphname, curr.rkg)
}
\arguments{
\item{rankings}{tables lrobject}

\item{slotname}{slot of the networks graphs_ggi to gene cell interaction and abs}

\item{graphname}{graph comparison name}

\item{curr.rkg}{ranking table}
}
\value{
list
}
\description{
Delta betweenness the most interactive gene (ligand or receptor)
}
