% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{comparative_pagerank}
\alias{comparative_pagerank}
\title{Ranking the most interactive gene (ligand or receptor)}
\usage{
comparative_pagerank(rankings, slotname, graphname, curr.rkg)
}
\arguments{
\item{ranking}{tables lrobject}

\item{slot}{slot of the networks graphs_ggi to gene cell interaction and abs}
}
\value{
list
}
\description{
Ranking the most interactive gene (ligand or receptor)
}
