% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparative_condition.R
\name{create_diff_table}
\alias{create_diff_table}
\title{Read the lrobject and generate the comparative tables}
\usage{
create_diff_table(data, out_path, comparison = NULL, score_col = "LRScore")
}
\arguments{
\item{data}{LRObj with single condition}

\item{out_path}{output path}
}
\value{
LRObject
}
\description{
Read the lrobject and generate the comparative tables
}
