% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparative_condition.R
\name{create_diff_table}
\alias{create_diff_table}
\title{Read the lrobject and generate the comparative tables}
\usage{
create_diff_table(data, out_path)
}
\arguments{
\item{lrpaths}{Paths of single condition LR data}

\item{sep}{character used on csv}
}
\value{
LRObject
}
\description{
Read the lrobject and generate the comparative tables
}
