% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Run all LR Downstream analysis}
\usage{
generate_report(
  LRpaths,
  genes,
  out_path,
  sep = ",",
  threshold = 50,
  colors = NULL,
  out_file = NULL
)
}
\arguments{
\item{LRpaths}{Paths of single condition LR data}

\item{sep}{character used on csv}
}
\value{
Rmarkdown report all objects from each step
}
\description{
This function loads the single conditions LR outputs and return the LR network based analysis.
It assumes that the table present the following columns Ligand, Ligand.Cluster, Receptor,Receptor.Cluster and MeanLR/another
measure
}
