% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Run all LR Downstream analysis}
\usage{
generate_report(
  lrpaths,
  genes = NULL,
  tf_genes = NULL,
  out_path,
  sep = ",",
  threshold = 0,
  colors = NULL,
  out_file = NULL,
  report = TRUE,
  output_fmt = "html_document",
  sel_columns = c("source", "target", "gene_A", "gene_B", "type_gene_A", "type_gene_B",
    "MeanLR"),
  org = "hsa",
  comparison = NULL,
  filtered_net = F,
  score_col = "LRScore",
  p_val = 0.05
)
}
\arguments{
\item{lrpaths}{Paths of single condition LR data}

\item{genes}{list of genes to be considered in the sankey plots}

\item{out_path}{output directory path}

\item{sep}{character used on csv}

\item{threshold}{percentage of edges to be pruned}

\item{colors}{celltypes colorscheme}

\item{out_file}{output file names}

\item{report}{decide if a report is generated or not}

\item{output_fmt}{rmarkdown render output format parameter}

\item{sel_columns}{columns from data}
}
\value{
Rmarkdown report all objects from each step
}
\description{
This function loads the single conditions LR outputs and return the LR network based analysis.
It assumes that the table present the following columns ('source','target','gene_A','gene_B','type_gene_A','type_gene_B','MeanLR')
measure
}
\examples{
paths <- c(
  "CTR" = system.file("extdata",
    "CTR_LR.csv",
    package = "CrossTalkeR"
  ),
  "EXP" = system.file("extdata",
    "EXP_LR.csv",
    package = "CrossTalkeR"
  )
)
output <- system.file("extdata", package = "CrossTalkeR")
genes <- c("TGFB1")
data <- generate_report(
  lrpaths = paths,
  genes = genes,
  out_path = paste0(output, "/"),
  threshold = 0,
  out_file = "report.html"
)

}
