% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{liana2CT}
\alias{liana2CT}
\title{Format liana2CT}
\usage{
liana2CT(data, source = "source", target = "target", gene_Ai, gene_Bi, measure)
}
\arguments{
\item{data}{datafromliana}

\item{source}{source cell population}

\item{target}{target cell population}

\item{gene_Ai}{source gene population}

\item{gene_Bi}{target gene population}

\item{measure}{measure to be considered}
}
\value{
tibble
}
\description{
Format liana2CT
}
