% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\docType{class}
\name{LRObj-class}
\alias{LRObj-class}
\alias{LRObject}
\title{Run and Generate all LR Downstream analysis}
\description{
This function loads the single conditions LR outputs and return the LR network based analysis.
It assumes that the table present the following columns Ligand, Ligand.Cluster, Receptor,Receptor.Cluster and MeanLR/another
measure
}
\section{Slots}{

\describe{
\item{\code{graphs}}{All Cell Cell Interaction Networks}

\item{\code{tables}}{All tables from single condition}

\item{\code{max_iter}}{Max meanLR from all}

\item{\code{max_nodes}}{All Celltype in the experiment}

\item{\code{coords}}{Cell Cell Interaction Plots}

\item{\code{colors}}{Cell type colors}
}}

