% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrobject.R
\docType{class}
\name{LRObj-class}
\alias{LRObj-class}
\alias{lrbject}
\title{Run and Generate all LR Downstream analysis}
\description{
This function loads the single conditions LR outputs and return the LR network
It assumes that the table present the following columns Ligand,
measure
}
\section{Slots}{

\describe{
\item{\code{graphs}}{All Cell Cell Interaction Networks}

\item{\code{tables}}{All tables from single condition}

\item{\code{max_iter}}{Max meanLR from all}

\item{\code{max_nodes}}{All Celltype in the experiment}

\item{\code{coords}}{Cell Cell Interaction Plots}

\item{\code{colors}}{Cell type colors}

\item{\code{rankings}}{Ranking of cells and Genes}

\item{\code{loadings}}{CCI values to remove multiple times genes}
}}

