% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{make_report}
\alias{make_report}
\title{Generate a report for given LRObj}
\usage{
make_report(
  genes = NULL,
  tf_genes = NULL,
  out_path,
  threshold = 0,
  colors = NULL,
  out_file = NULL,
  output_fmt = "html_document",
  LRObj = NULL,
  sel_columns = c("source", "target", "gene_A", "gene_B", "type_gene_A", "type_gene_B",
    "MeanLR")
)
}
\arguments{
\item{genes}{list of genes to be considered in the sankey plots}

\item{out_path}{output directory path}

\item{threshold}{percentage of edges to be pruned}

\item{colors}{celltypes colorscheme}

\item{out_file}{output file names}

\item{output_fmt}{rmarkdown render output format parameter}

\item{LRObj}{rmarkdown render output format parameter}

\item{sel_columns}{columns from data}
}
\value{
Rmarkdown report all objects from each step
}
\description{
Generate a report for given LRObj
}
\examples{
paths <- c('CTR' = system.file("extdata",
                              "CTR_LR.csv",
                              package = "CrossTalkeR"),
          'EXP' = system.file("extdata",
                              "EXP_LR.csv",
                              package = "CrossTalkeR"))
output_path =  system.file("extdata", package = "CrossTalkeR")
genes <- c('TGFB1')
data <- generate_report(lrpaths = paths,
                       genes = genes,
                       out_path = output_path,
                       threshold = 0,
                         out_file = "report.html")

}
