% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_articulation}
\alias{plot_articulation}
\title{This function do a ggi plot and articulation}
\usage{
plot_articulation(graph, color)
}
\arguments{
\item{graph}{graph}

\item{color}{cluster color}
}
\value{
R default plot
}
\description{
This function do a ggi plot and articulation
}
