% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ggi}
\alias{plot_ggi}
\title{This function do a ggi plot and higest degree nodes}
\usage{
plot_ggi(graph, color, name)
}
\arguments{
\item{graph}{graph}

\item{color}{cluster color}

\item{name}{plot header}
}
\value{
R default plot
}
\description{
This function do a ggi plot and higest degree nodes
}
\examples{
paths <- c('CTR' = system.file("extdata",
                              "ctr_nils_bm_human.csv",
                              package = "CrossTalkeR"),
          'EXP' = system.file("extdata",
                              "exp_nils_bm_human.csv",
                              package = "CrossTalkeR"))
output =  system.file("extdata", package = "CrossTalkeR")
genes <- c('TGFB1')
data <- generate_report(paths,
                       genes,
                       out_path=paste0(output,'/'),
                       threshold=0,
                       out_file = 'vignettes_example.html',
                       output_fmt = "html_document",
                       report = FALSE)
plot_ggi(graph = data@graphs_ggi$EXP_x_CTR,
        color = data@colors,name="EXP_x_CTR")
}
