% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pca_LR_comparative}
\alias{plot_pca_LR_comparative}
\title{This function is a proxy to the PCA plot in comparative conditions}
\usage{
plot_pca_LR_comparative(
  lrobj_tblPCA,
  dims = c(1, 2),
  ret = F,
  ggi = TRUE,
  include_tf = TRUE,
  gene_types = "all"
)
}
\arguments{
\item{dims}{PCA dims}

\item{ret}{return plot}

\item{ggi}{GGI mode}

\item{lrobj_tbl}{LRobject table with all data}

\item{curr}{table entry}
}
\value{
R default plot
}
\description{
This function is a proxy to the PCA plot in comparative conditions
}
\examples{
paths <- c('CTR' = system.file("extdata",
                              "ctr_nils_bm_human_newformat.csv",
                              package = "CrossTalkeR"),
          'EXP' = system.file("extdata",
                              "exp_nils_bm_human_newformat.csv",
                              package = "CrossTalkeR"))
output =  system.file("extdata", package = "CrossTalkeR")
genes <- c('TGFB1')

data <- generate_report(paths,
                       genes,
                       out_path=paste0(output,'/'),
                       threshold=0,
                       out_file = 'vignettes_example.html',
                       output_fmt = "html_document",
                       report = FALSE)
}
