% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sankey}
\alias{plot_sankey}
\title{This function selected genes sankey plot}
\usage{
plot_sankey(
  LRObj_tbl,
  target = NULL,
  Ligand.Cluster = NULL,
  Receptor.Cluster = NULL,
  pltname = NULL
)
}
\arguments{
\item{LRObj_tbl}{LRobject with all data}

\item{target}{gene}

\item{lcls}{Ligand Clusters}

\item{rcls}{Receptor Clusters}
}
\value{
R default plot
}
\description{
This function selected genes sankey plot
}
