% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sankey}
\alias{plot_sankey}
\title{This function selected genes sankey plot}
\usage{
plot_sankey(
  lrobj_tbl,
  target = NULL,
  ligand_cluster = NULL,
  receptor_cluster = NULL,
  pltname = NULL
)
}
\arguments{
\item{lrobj_tbl}{LRobject with all data}

\item{target}{gene}

\item{lcls}{Ligand Clusters}

\item{rcls}{Receptor Clusters}
}
\value{
R default plot
}
\description{
This function selected genes sankey plot
}
