% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sankey}
\alias{plot_sankey}
\title{This function selected genes sankey plot}
\usage{
plot_sankey(
  lrobj_tbl,
  target = NULL,
  ligand_cluster = NULL,
  receptor_cluster = NULL,
  plt_name = NULL,
  threshold = 50
)
}
\arguments{
\item{lrobj_tbl}{LRobject table with all data}

\item{target}{gene}

\item{ligand_cluster}{Ligand Clusters}

\item{receptor_cluster}{Receptor Clusters}

\item{plt_name}{plot title}
}
\value{
R default plot
}
\description{
This function selected genes sankey plot
}
