% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_signedbar_ggi}
\alias{plot_signedbar_ggi}
\title{This function signed sending and receiving barplot (GGI)}
\usage{
plot_signedbar_ggi(all_data, curr)
}
\arguments{
\item{all_data}{LRobject}

\item{curr}{condition}
}
\value{
R default plot
}
\description{
This function signed sending and receiving barplot (GGI)
}
