% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ranking}
\alias{ranking}
\title{Ranking the most interactive gene (ligand or receptor)}
\usage{
ranking(data, out_path, slot = "graphs_ggi")
}
\arguments{
\item{data}{lrobject}

\item{out}{path to save the lrobject with ranking}
}
\value{
list
}
\description{
Ranking the most interactive gene (ligand or receptor)
}
